avm2: Alternative implementation of XML hasOwnProperty @attribute
